moved module detection code out of global space, moved errors to separate file
