+++ screep
@@ -6,13 +6,16 @@ enc = ["b64"]
 
 dataz = "very secret and private message"
 
-print Exfil.encoders()
+print Exfil.list_encoders()
 
 t = Exfil(mod, enc)
 t.set_channel_params({'sending': {'filename': 'test.txt'}})
 t.set_channel_params({'receiving': {'filename': 'test.txt'}})
 t.set_encoder_params(enc[0], {})
+
 print t.channel_config()
 print t.encoder_config(enc[0])
+
 t.send(dataz)
+
 print t.receive()
